# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: FoodFlow API
  description: |
    FoodFlow -- Food Delivery Platform REST API.

    **Base path:** `/api`. The production origin is the verified Vercel API alias.

    **Conventions:**
    - All monetary values are in **VND**, represented as integers.
    - All datetime values use ISO 8601 `date-time` with timezone.
    - Successful web/admin/restaurant endpoints return `{ success: true, data, meta? }`.
    - List endpoints put the collection in `data` and pagination in `meta`.
    - Errors use the shared `ProblemDetail` schema (RFC 7807) and are not success-wrapped.
    - Idempotency is supported via `Idempotency-Key` header (UUID v4).
    - Inbound HMAC: SePay webhooks are verified with `x-sepay-signature`.
    - Outbound HMAC: generic outbound webhooks use `X-Signature-SHA256` when configured.

    **Versioning:** Batch 4 currently uses the `/api` base without a `/v1` prefix. A future version prefix requires an updated server contract and migration policy.
  version: "1.0.0"
  contact:
    name: FoodFlow Team
    email: support@foodflow.vn
servers:
  - url: https://replace-with-foodflow-api.vercel.app/api
    description: Production placeholder; replace with the verified Vercel API alias before publishing external client configuration
  - url: http://localhost:3001/api
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer access token. Obtain via `POST /auth/login`; the signing algorithm follows the active JWT cutover phase.
    cronBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Exact server-side CRON_SECRET value. Never expose this credential to browser or mobile clients.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Page number (1-indexed)
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Items per page
...
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:    { type: string, description: "URI to problem type documentation" }
        title:   { type: string, description: "Short human-readable summary" }
        status:  { type: integer, description: "HTTP status code" }
        detail:  { type: string, description: "Human-readable explanation" }
        instance: { type: string, description: "URI to the specific occurrence" }
        code:    { type: string, description: "Application-specific error code (e.g. AUTH_INVALID_CREDENTIALS)" }
      required: [type, title, status, detail]
      example:
        type: "about:blank"
        title: "Validation Error"
        status: 422
        detail: "Email is not valid"
        code: "VALIDATION_INVALID_EMAIL"
...
    RestaurantDetail:
      allOf:
        - $ref: "#/components/schemas/RestaurantSummary"
        - type: object
          properties:
            description: { type: string }
            addressLine: { type: string }
            city:        { type: string }
            district:    { type: string }
            phone:       { type: string }
            openingHours:
              type: array
              items: { $ref: "#/components/schemas/RestaurantOpeningHour" }
...
    MenuCategory:
      type: object
      properties:
        id:       { type: string, format: uuid }
        name:     { type: string }
        sortOrder: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/MenuItem" }
      required: [id, name, sortOrder, items]
...
    RestaurantMenuAvailabilityRequest:
      type: object
      properties:
        available: { type: boolean }
      required: [available]
...
    Promotion:
      type: object
      properties:
        id:            { type: string, format: uuid }
        code:          { type: string }
        name:          { type: string, maxLength: 150 }
        description:   { type: string }
        type:          { type: string, enum: [percentage, fixed, free_delivery, bogo, combo] }
        value:         { type: integer, description: "Percent (if percentage) or VND (if fixed)" }
        minOrderAmount: { type: integer, description: "Minimum order amount in VND" }
        maxDiscount:   { type: integer, description: "Max discount cap in VND" }
        usageLimit:    { type: integer }
        usageCount:    { type: integer }
        maxPerUser:    { type: integer, minimum: 1 }
        targeting:     { type: object, additionalProperties: true }
        status:        { type: string, enum: [draft, scheduled, active, paused, expired, archived] }
        startsAt:      { type: string, format: date-time }
        expiresAt:     { type: string, format: date-time }
        isActive:      { type: boolean }
        createdAt:     { type: string, format: date-time }
      required: [id, code, type, value, usageLimit, startsAt, expiresAt, isActive]
      example:
        id: "88888888-8888-4888-8888-888888888888"
        code: "WELCOME2026"
        type: "percentage"
        value: 20
        minOrderAmount: 100000
        maxDiscount: 50000
        usageLimit: 500
        usageCount: 23
        startsAt: "2026-06-01T00:00:00Z"
        expiresAt: "2026-07-01T00:00:00Z"
        isActive: true
        createdAt: "2026-06-01T00:00:00Z"
...
    PaginatedReviews:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Review" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
      required: [data, meta]
paths:
...
  /restaurants/{restaurantId}:
    get:
      tags: [restaurants]
      summary: Get restaurant detail
      description: Returns public detail for one active restaurant.
      operationId: restaurants_getDetail
      x-foo-actor: public
      security: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Restaurant detail", content: { application/json: { schema: { $ref: "#/components/schemas/RestaurantDetail" } } } }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /restaurants/{restaurantId}/menu:
    get:
      tags: [restaurants]
      summary: Get restaurant menu (categories + items)
      description: Returns visible menu categories and available items for one active restaurant.
      operationId: restaurants_getMenu
      x-foo-actor: public
      security: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Restaurant menu", content: { application/json: { schema: { type: object, properties: { categories: { type: array, items: { $ref: "#/components/schemas/MenuCategory" } } } } } } }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /restaurants/{restaurantId}/reviews:
    get:
      tags: [restaurants]
      summary: Get restaurant reviews
      description: Lists public customer reviews for one restaurant with pagination.
      operationId: restaurants_getReviews
      x-foo-actor: public
      security: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200": { description: "Restaurant reviews", content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedReviews" } } } }

  # ═══════════════════════════════════════════════
  # MENU (restaurant owner)
  # ═══════════════════════════════════════════════
...
  /orders/{orderId}:
    get:
      tags: [orders]
      summary: Get order detail
      description: Returns one order when it belongs to the authenticated customer.
      operationId: orders_getDetail
      x-foo-actor: user
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Order detail", content: { application/json: { schema: { $ref: "#/components/schemas/OrderDetail" } } } }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /promotions/{promotionCode}:
    get:
      tags: [promotions]
      summary: Get promotion by code
      description: Returns a promotion by code when it is visible and accessible to the authenticated user.
      operationId: promotions_getByCode
      x-foo-actor: user
      parameters:
        - name: promotionCode
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "Promotion detail", content: { application/json: { schema: { $ref: "#/components/schemas/Promotion" } } } }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }

  # ═══════════════════════════════════════════════
  # NOTIFICATIONS
  # ═══════════════════════════════════════════════
...
  /admin/users/{userId}:
    get:
      tags: [analytics]
      summary: Get one user for admin review
      description: Returns one user with customer, driver, and restaurant profile details. Password hashes are omitted by the backend.
      operationId: admin_getUser
      x-foo-actor: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "User detail", content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "404": { description: "User not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/users/{userId}/wallet:
    get:
      tags: [analytics]
      summary: Get one user's wallet ledger
      description: Returns the confirmed balance derived from real wallet transactions and the latest wallet ledger rows.
      operationId: admin_getUserWallet
      x-foo-actor: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: "Wallet ledger"
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance: { type: number }
                  transactions:
                    type: array
                    items: { type: object, additionalProperties: true }
                required: [balance, transactions]
        "404": { description: "User not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/users/{userId}/vouchers:
    get:
      tags: [analytics]
      summary: Get voucher and promotion usage for one user
      description: Returns currently usable persisted promotions and real promotion usage history for the user. Empty arrays mean no matching persisted promotions/usages; no placeholder voucher rows are synthesized.
      operationId: admin_getUserVouchers
      x-foo-actor: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: "User voucher and promotion usage summary"
          content:
            application/json:
              schema:
                type: object
                properties:
                  owned:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        promotionId: { type: string, format: uuid }
                        code: { type: string }
                        description: { type: string }
                        discountType: { type: string, enum: [percentage, fixed, free_delivery, bogo, combo] }
                        discountValue: { type: number }
                        minOrder: { type: number }
                        maxDiscount: { type: number }
                        validUntil: { type: string, format: date-time }
                        usedAt: { type: string, format: date-time, nullable: true }
                        orderCode: { type: string, nullable: true }
                      required: [id, promotionId, code, description, discountType, discountValue, minOrder, maxDiscount, validUntil, usedAt]
                  used:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        promotionId: { type: string, format: uuid }
                        code: { type: string }
                        description: { type: string }
                        discountType: { type: string, enum: [percentage, fixed, free_delivery, bogo, combo] }
                        discountValue: { type: number }
                        minOrder: { type: number }
                        maxDiscount: { type: number }
                        validUntil: { type: string, format: date-time }
                        usedAt: { type: string, format: date-time, nullable: true }
                        orderCode: { type: string, nullable: true }
                      required: [id, promotionId, code, description, discountType, discountValue, minOrder, maxDiscount, validUntil, usedAt]
                  totalSaved: { type: number }
                required: [owned, used, totalSaved]
        "404": { description: "User not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/users/{userId}/refunds:
    get:
      tags: [analytics]
      summary: Get refunded payments for one user
      description: Returns real refunded payment rows linked to the user's customer orders.
      operationId: admin_getUserRefunds
      x-foo-actor: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Refunded payments", content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
        "404": { description: "User not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/users/{userId}/kyc:
    get:
      tags: [drivers]
      summary: Get driver KYC submissions for one user
      description: Returns an explicit unavailable reason for non-driver users or real driver KYC submissions with five-minute signed private-document read URLs for authorized admin review. Raw object keys are never returned.
      operationId: admin_getUserKyc
      x-foo-actor: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: "KYC payload"
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    additionalProperties: false
                    properties:
                      available: { type: boolean, const: false }
                      reason: { type: string, enum: [NOT_A_DRIVER] }
                    required: [available, reason]
                  - type: object
                    additionalProperties: false
                    properties:
                      available: { type: boolean, const: true }
                      submissions:
                        type: array
                        items: { $ref: "#/components/schemas/AdminKycSubmission" }
                    required: [available, submissions]
        "404": { description: "User not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/users/{userId}/kyc/review:
    post:
      tags: [drivers]
      summary: Review one driver KYC submission
      description: Approves or rejects a real KYC submission and records the reviewing admin.
      operationId: admin_reviewUserKyc
      x-foo-actor: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KycReviewRequest" }
      responses:
        "201": { description: "KYC review saved without private object keys", content: { application/json: { schema: { $ref: "#/components/schemas/AdminKycReviewResponse" } } } }
        "400": { description: "Rejection reason or private document manifest is invalid", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
        "409": { description: "KYC submission was already reviewed", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
        "404": { description: "KYC submission not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /admin/restaurants/{restaurantId}:
    get:
      tags: [analytics]
      summary: Get one restaurant for admin review
      description: Returns one restaurant with opening hours and owner profile relations for admin detail pages.
      operationId: admin_getRestaurant
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Restaurant detail", content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/overview:
    get:
      tags: [analytics]
      summary: Get restaurant operational overview
      description: Returns real aggregate order, revenue, rating, menu, and staff counts for one restaurant.
      operationId: admin_getRestaurantOverview
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: "Restaurant overview"
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOrders: { type: integer, minimum: 0 }
                  completedOrders: { type: integer, minimum: 0 }
                  revenue: { type: number, minimum: 0 }
                  rating: { type: number, minimum: 0, maximum: 5 }
                  reviewCount: { type: integer, minimum: 0 }
                  menuItems: { type: integer, minimum: 0 }
                  activeStaff: { type: integer, minimum: 0 }
                required: [totalOrders, completedOrders, revenue, rating, reviewCount, menuItems, activeStaff]
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/menu:
    get:
      tags: [restaurants]
      summary: Get restaurant menu for admin review
      description: Returns real menu categories, items, options, and option values for one restaurant.
      operationId: admin_getRestaurantMenu
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Restaurant menu", content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/menu/{itemId}:
    patch:
      tags: [restaurants]
      summary: Update one restaurant menu item's availability
      description: Updates the real menu item availability after verifying it belongs to the selected restaurant.
      operationId: admin_updateRestaurantMenuItem
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RestaurantMenuAvailabilityRequest" }
      responses:
        "200": { description: "Menu item updated", content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "404": { description: "Menu item not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/orders:
    get:
      tags: [analytics]
      summary: List orders for one restaurant as admin
      description: Returns real orders for one restaurant with pagination.
      operationId: admin_getRestaurantOrders
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200": { description: "Restaurant orders", content: { application/json: { schema: { type: object, properties: { orders: { type: array, items: { $ref: "#/components/schemas/OrderDetail" } }, meta: { $ref: "#/components/schemas/PaginationMeta" } }, required: [orders, meta] } } } }
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/reviews:
    get:
      tags: [restaurants]
      summary: List reviews for one restaurant as admin
      description: Returns real review rows for one restaurant with pagination.
      operationId: admin_getRestaurantReviews
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200": { description: "Restaurant reviews", content: { application/json: { schema: { type: object, properties: { reviews: { type: array, items: { type: object, additionalProperties: true } }, meta: { $ref: "#/components/schemas/PaginationMeta" } }, required: [reviews, meta] } } } }
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/finance:
    get:
      tags: [analytics]
      summary: Get restaurant finance summary
      description: Returns real gross, commission, net, and completed order count for one restaurant.
      operationId: admin_getRestaurantFinance
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: "Restaurant finance"
          content:
            application/json:
              schema:
                type: object
                properties:
                  gross: { type: number, minimum: 0 }
                  commission: { type: number, minimum: 0 }
                  net: { type: number }
                  orderCount: { type: integer, minimum: 0 }
                required: [gross, commission, net, orderCount]
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/approve:
    post:
      tags: [restaurants]
      summary: Approve one restaurant
      description: Approves a pending restaurant, records admin notes, sets it active, and emits an audit log.
      operationId: admin_approveRestaurant
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RestaurantReviewRequest" }
      responses:
        "201": { description: "Restaurant approved", content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/restaurants/{restaurantId}/reject:
    post:
      tags: [restaurants]
      summary: Reject one restaurant
      description: Rejects a pending restaurant, records admin notes or reason, and emits an audit log.
      operationId: admin_rejectRestaurant
      x-foo-actor: admin
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RestaurantReviewRequest" }
      responses:
        "201": { description: "Restaurant rejected", content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "404": { description: "Restaurant not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /admin/promotions/{promotionId}:
    get:
      tags: [promotions]
      summary: Get one promotion (admin)
      description: Returns one platform promotion with its promotion items for the admin detail page.
      operationId: admin_getPromotion
      x-foo-actor: admin
      parameters:
        - name: promotionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Promotion", content: { application/json: { schema: { $ref: "#/components/schemas/Promotion" } } } }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
    patch:
      tags: [promotions]
      summary: Update a promotion (admin)
      description: Updates a platform promotion selected by an admin.
      operationId: admin_updatePromotion
      x-foo-actor: admin
      parameters:
        - name: promotionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePromotionRequest" }
      responses:
        "200": { description: "Promotion updated", content: { application/json: { schema: { $ref: "#/components/schemas/Promotion" } } } }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
    delete:
      tags: [promotions]
      summary: Delete a promotion (admin)
      description: Deletes a platform promotion selected by an admin when backend validation allows removal.
      operationId: admin_deletePromotion
      x-foo-actor: admin
      parameters:
        - name: promotionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Promotion deleted" }
        "404": { description: "Not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /admin/promotions/{promotionId}/analytics:
    get:
      tags: [promotions]
      summary: Get admin promotion analytics
      description: Returns real promotion redemption count, GMV, discount cost, and ROI from promotion usage rows.
      operationId: admin_getPromotionAnalytics
      x-foo-actor: admin
      parameters:
        - name: promotionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Promotion analytics", content: { application/json: { schema: { $ref: "#/components/schemas/AdminPromotionAnalytics" } } } }
        "404": { description: "Promotion not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
...
  /admin/orders/{orderId}:
    get:
      tags: [analytics]
      summary: Get one order for admin review
      description: Returns a serialized order with customer, restaurant, driver, address, items, payment, delivery task, and status history details.
      operationId: admin_getOrder
      x-foo-actor: admin
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Order detail", content: { application/json: { schema: { $ref: "#/components/schemas/OrderDetail" } } } }
        "404": { description: "Order not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
  /admin/orders/{orderId}/status:
    patch:
      tags: [analytics]
      summary: Update one order status as admin
      description: Applies an admin order status transition through the backend order state machine and audit logging.
      operationId: admin_updateOrderStatus
      x-foo-actor: admin
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminOrderStatusRequest" }
      responses:
        "200": { description: "Order status updated", content: { application/json: { schema: { $ref: "#/components/schemas/OrderDetail" } } } }
        "400": { description: "Invalid transition", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }
        "404": { description: "Order not found", content: { application/json: { schema: { $ref: "#/components/schemas/ProblemDetail" } } } }